python/requirements.txt
PyPDF2
ghostscript
pytesseract
tabula-py